[td] Trigger and run pipelines across projects (#4186)
* [td] Trigger and run pipelines across projects

* fix

* fix tests

* fix tests

* fix

* fix

* fix

* fix

* fix

* hooks run cross project

* fix

* fix

* [td] Build FRONTEND assets

* fix tests

* fix tests

* fix tests

* fix

* Revert "[td] Build FRONTEND assets"

This reverts commit 2e176e74943ede76eeea5d65ad91e85852227eb3.

* pipeline resource

* pipeline runs

* block

* block

* pipeline hooks

* pipeline

* variable manager

* platform

* platform utils

* platform utils

* hook

* hook

* global hooks

* pipeline scheduler

* pipeline scheduler

* fix

* fix

* fix logs

* fix logs

* fix logs

* fix logs

* fix

* fix

* fix

* fix

* fix

+++ mage_ai/tests/shared/mixins.py
@@ -29,10 +29,12 @@ from mage_ai.data_preparation.models.global_hooks.predicates import (
     PredicateValueType,
 )
 from mage_ai.settings.platform import (
+    build_repo_path_for_all_projects,
     local_platform_settings_full_path,
     platform_settings_full_path,
 )
 from mage_ai.settings.repo import get_repo_path
+from mage_ai.settings.utils import base_repo_path
 from mage_ai.shared.hash import merge_dict
 from mage_ai.shared.io import safe_write
 from mage_ai.tests.api.operations.test_base import BaseApiTestCase
@@ -368,9 +370,14 @@ class ProjectPlatformMixin(AsyncDBTestCase):
 
     @classmethod
     def tearDownClass(self):
-        if os.path.exists(platform_settings_full_path()):
+        if os.path.exists(platform_settings_full_path()) and \
+                base_repo_path() != platform_settings_full_path():
+
             os.remove(platform_settings_full_path())
-        if os.path.exists(local_platform_settings_full_path()):
+
+        if os.path.exists(local_platform_settings_full_path()) and \
+                base_repo_path() != local_platform_settings_full_path():
+
             os.remove(local_platform_settings_full_path())
 
         try:
@@ -387,6 +394,7 @@ class ProjectPlatformMixin(AsyncDBTestCase):
                 self.repo_path,
                 return_blocks=True,
             )
+            self.repo_paths = build_repo_path_for_all_projects(mage_projects_only=True)
 
     def teardown_final(self):
         with patch('mage_ai.settings.platform.project_platform_activated', lambda: True):
